requirements.txt
asgiref==3.3.1
certifi==2020.12.5
Django==3.1.7
django-filter==2.4.0
djangorestframework==3.12.2
httpie-jwt-auth==0.4.0
pytz==2021.1
sqlparse==0.4.1
unrpa==2.3.0
